Move completion classes to intellisense package
Refactored auto-completion related classes from pseudopad.editor.completion to pseudopad.editor.intellisense for improved clarity and organization. Updated imports in FileTabPane accordingly.
